browser/components/payments/test/mochitest/test_payment_details_item.html

<!DOCTYPE HTML>
<html>
<!--
Test the payment-details-item component
-->
<head>
  <meta charset="utf-8">
  <title>Test the payment-details-item component</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="application/javascript" src="/tests/SimpleTest/AddTask.js"></script>
  <script src="payments_common.js"></script>
  <script src="../../res/vendor/custom-elements.min.js"></script>

  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
  <p id="display">
    <payment-details-item id="item1"></payment-details-item>
    <payment-details-item id="item2" label="Some item" amount-value="2" amount-currency="USD"></payment-details-item>
  </p>
<div id="content" style="display: none">

</div>
<pre id="test">
</pre>
<script type="module">
/** Test the payment-details-item component **/

import "../../res/components/payment-details-item.js";

let item1 = document.getElementById("item1");
let item2 = document.getElementById("item2");

add_task(async function test_no_value() {
  ok(item1, "item1 exists");
  is(item1.textContent, "", "Initially empty");

  item1.label = "New label";
  await asyncElementRendered();
  is(item1.getAttribute("label"), "New label", "Check @label");
  ok(!item1.hasAttribute("amount-value"), "Check @amount-value");
  ok(!item1.hasAttribute("amount-currency"), "Check @amount-currency");
  is(item1.label, "New label", "Check .label");
  is(item1.amountValue, null, "Check .amountValue");
  is(item1.amountCurrency, null, "Check .amountCurrency");

  item1.label = null;
  await asyncElementRendered();
  ok(!item1.hasAttribute("label"), "Setting to null should remove @label");
  is(item1.textContent, "", "Becomes empty when label is removed");
});

add_task(async function test_initial_attribute_values() {
  is(item2.label, "Some item", "Check .label");
  is(item2.amountValue, "2", "Check .amountValue");
  is(item2.amountCurrency, "USD", "Check .amountCurrency");
});

add_task(async function test_templating() {
  ok(item2.querySelector("currency-amount"), "creates currency-amount component");
  ok(item2.querySelector(".label"), "creates label");
});

</script>

</body>
</html>
